Optimize Fumadocs source loading
