+++ app/(home)/blog/page.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import { blog } from '@/lib/source';
+import { blog } from '@/lib/source/posts';
 import Image from 'next/image';
 import { formatDate } from '@/lib/utils';
