+++ lib/source/docs.ts
@@ -1,10 +1,8 @@
-import { docs, blogPosts, newsPosts } from 'fumadocs-mdx:collections/server';
+import { docs } from 'fumadocs-mdx:collections/dynamic';
 import { type InferPageType, loader } from 'fumadocs-core/source';
 import { lucideIconsPlugin } from 'fumadocs-core/source/lucide-icons';
-import { toFumadocsSource } from 'fumadocs-mdx/runtime/server';
 import { i18n } from '@/lib/i18n';
 
-// See https://fumadocs.dev/docs/headless/source-api for more info
 export const source = loader({
   baseUrl: '/docs',
   source: docs.toFumadocsSource(),
@@ -20,13 +18,3 @@ export function getPageImage(page: InferPageType<typeof source>) {
     url: `/og/docs/${segments.join('/')}`,
   };
 }
-
-export const blog = loader({
-  baseUrl: '/blog',
-  source: toFumadocsSource(blogPosts, []),
-});
-
-export const news = loader({
-  baseUrl: '/news',
-  source: toFumadocsSource(newsPosts, []),
-});
